primitives: Move block test to tests
This test uses the `include_bytes!` macro to include data from a file
that lives in `tests` but we exclude that directory when publishing.
Resulting in published source code having broken unit tests.

Move the test to `tests/`.
